[프로젝트 설정] 프로젝트 설정 project_setting으로 구조 변경 및 공수 삭제 기능 추가
1. project와 man_hour로 나누어져 있던 기능들을 project_setting으로 통합
2. 공수 삭제 기능 추가
