+++ csm-api/entity/project.go
@@ -86,15 +86,3 @@ type ReqProject struct {
 	WorkRate  null.Int    `json:"work_rate" db:"WORK_RATE"`
 	Base
 }
-
-type ProjectSetting struct {
-	Jno         null.Int    `json:"jno" db:"JNO"`
-	InTime      null.Time   `json:"in_time" db:"IN_TIME"`           // 출근시간
-	OutTime     null.Time   `json:"out_time" db:"OUT_TIME"`         // 퇴근시간
-	RespiteTime null.Int    `json:"respite_time" db:"RESPITE_TIME"` // 출/퇴근 유예시간(분)
-	CancelCode  null.String `json:"cancel_code" db:"CANCEL_CODE"`   // 마감취소가능기한 CODE
-	ManHours    *ManHours   `json:"man_hours"`
-	Base
-}
-
-type ProjectSettings []*ProjectSetting
