+++ csm-api/mux.go
@@ -64,21 +64,21 @@ func newMux(ctx context.Context, safeDb *sqlx.DB, timesheetDb *sqlx.DB) (http.Ha
 		//		next.ServeHTTP(w, r)
 		//	})
 		//})
-		router.Mount("/api", route.ApiRoute(apiCfg, safeDb, &r))               // api
-		router.Mount("/excel", route.ExcelRoute(safeDb, &r))                   // 엑셀
-		router.Mount("/project", route.ProjectRoute(safeDb, &r))               // 프로젝트
-		router.Mount("/organization", route.OrganiztionRoute(timesheetDb, &r)) // 조직도
-		router.Mount("/site", route.SiteRoute(safeDb, &r, apiCfg))             // 현장
-		router.Mount("/worker", route.WorkerRoute(safeDb, &r))                 // 근로자
-		router.Mount("/compare", route.CompareRoute(safeDb, &r))               // 일일 근로자 비교
-		router.Mount("/deadline", route.DeadlineRoute(safeDb, &r))             // 일일마감
-		router.Mount("/equip", route.EquipRoute(safeDb, &r))                   // 장비 (임시)
-		router.Mount("/device", route.DeviceRoute(safeDb, &r))                 // 근태인식기
-		router.Mount("/company", route.CompanyRoute(safeDb, timesheetDb, &r))  // 협력업체
-		router.Mount("/schedule", route.ScheduleRoute(safeDb, &r))             // 일정관리
-		router.Mount("/notice", route.NoticeRoute(safeDb, &r))                 // 공지사항
-		router.Mount("/code", route.CodeRoute(safeDb, &r))                     // 코드
-
+		router.Mount("/api", route.ApiRoute(apiCfg, safeDb, &r))                // api
+		router.Mount("/excel", route.ExcelRoute(safeDb, &r))                    // 엑셀
+		router.Mount("/project", route.ProjectRoute(safeDb, &r))                // 프로젝트
+		router.Mount("/organization", route.OrganiztionRoute(timesheetDb, &r))  // 조직도
+		router.Mount("/site", route.SiteRoute(safeDb, &r, apiCfg))              // 현장
+		router.Mount("/worker", route.WorkerRoute(safeDb, &r))                  // 근로자
+		router.Mount("/compare", route.CompareRoute(safeDb, &r))                // 일일 근로자 비교
+		router.Mount("/deadline", route.DeadlineRoute(safeDb, &r))              // 일일마감
+		router.Mount("/equip", route.EquipRoute(safeDb, &r))                    // 장비 (임시)
+		router.Mount("/device", route.DeviceRoute(safeDb, &r))                  // 근태인식기
+		router.Mount("/company", route.CompanyRoute(safeDb, timesheetDb, &r))   // 협력업체
+		router.Mount("/schedule", route.ScheduleRoute(safeDb, &r))              // 일정관리
+		router.Mount("/notice", route.NoticeRoute(safeDb, &r))                  // 공지사항
+		router.Mount("/code", route.CodeRoute(safeDb, &r))                      // 코드
+		router.Mount("/project-setting", route.ProjectSettingRoute(safeDb, &r)) // 프로젝트 설정
 	})
 
 	return c.Handler(mux), nil
