+++ csm-api/route/route_project.go
@@ -17,11 +17,11 @@ func ProjectRoute(safeDB *sqlx.DB, r *store.Repository) chi.Router {
 			SafeTDB:   safeDB,
 			Store:     r,
 			UserStore: r,
-			ManHourService: &service.ServiceManHour{
-				SafeDB:  safeDB,
-				SafeTDB: safeDB,
-				Store:   r,
-			},
+			//ManHourService: &service.ServiceManHour{
+			//	SafeDB:  safeDB,
+			//	SafeTDB: safeDB,
+			//	Store:   r,
+			//},
 		},
 	}
 
@@ -32,10 +32,8 @@ func ProjectRoute(safeDB *sqlx.DB, r *store.Repository) chi.Router {
 	router.Get("/my-org/{uno}", projectHandler.MyOrgList)          // 본인이 속한 조직도의 프로젝트 조회
 	router.Get("/my-job_name/{uno}", projectHandler.MyJobNameList) // 본인이 속한 프로젝트 이름 목록
 	router.Get("/non-reg", projectHandler.NonRegList)              // 현장근태 사용되지 않은 프로젝트
-	router.Get("/setting/{jno}", projectHandler.JobSetting)        // 프로젝트 기본 설정 정보 조회
 	router.Get("/project-by-site", projectHandler.ProjectBySite)   // 현장별 프로젝트 조회
 	router.Post("/", projectHandler.Add)                           // 추가
-	router.Post("/setting", projectHandler.MergeJobSetting)        // 프로젝트 기본 정보 추가 및 수정
 	router.Put("/default", projectHandler.ModifyDefault)           // 현장 기본 프로젝트 변경
 	router.Put("/use", projectHandler.ModifyIsUse)                 // 현장 프로젝트 사용여부 변경
 	router.Delete("/{sno}/{jno}", projectHandler.Remove)           // 현장 프로젝트 삭제
